feat: 新增下载管理功能，包含下载服务、界面及通知支持
Former-commit-id: 3a40a074df467f7b08a4430d95816237034b92e7

+++ app/src/main/java/com/imcys/bilibilias/dwonload/AppDownloadTask.kt
@@ -12,6 +12,6 @@ data class AppDownloadTask(
     val downloadSubTasks: List<DownloadSubTask>,
     val downloadStage: DownloadStage = DownloadStage.DOWNLOAD,
     val cover: String?,
-    val progress: Double = 0.0,
+    val progress: Float = 0.0f,
     val downloadState: DownloadState = DownloadState.WAITING,
 )
